+++ BOCCHI.MobFarmer/Services/IMobFarmer.cs
@@ -17,6 +17,7 @@ namespace BOCCHI.MobFarmer.Data
     {
         None = 0,
         Pots,
+        TreasureSight,
         TreasureHunt,
         CrystalBuffs,
     }